% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjLF.R
\name{adjLF}
\alias{adjLF}
\title{Make Init L and F on similar scale}
\usage{
adjLF(L, FF)
}
\arguments{
\item{L}{N by K loading matrix}

\item{FF}{K by p factor matrix}
}
\description{
Make Init L and F on similar scale
}
\details{
LF = L \times S_l^{-1} \times S_l \times S_f \times S_f^{-1} \times F
}
