% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSM.R
\name{bmsm_control_default}
\alias{bmsm_control_default}
\title{Default control list of BMSM}
\usage{
bmsm_control_default()
}
\arguments{
\item{reflect:}{whether reflect x to make it have length of power of 2.}

\item{shape:}{shape of Beta prior parameters, default to be 100, 50, 20, 10, 5, 2, 1.}

\item{point_mass:}{whether put a point mass at 1/2 when performing EB estimate of binomial probability.}

\item{nullweight:}{nullweight to induce more smoothness.}

\item{control:}{controls of mixSQP}

\item{Elogl:}{whether calculate and return E(log(lambda))}
}
\description{
Default control list of BMSM
}
