% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_log.R
\name{calc_ebpmf_log_obj}
\alias{calc_ebpmf_log_obj}
\title{Calc elbo}
\usage{
calc_ebpmf_log_obj(
  n,
  p,
  sym,
  ssexp,
  slogv,
  sv,
  sigma2,
  R2,
  KL_LF,
  const,
  ss,
  a0,
  b0
)
}
\description{
Calc elbo
}
