% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_identity.R
\name{Calc_EZ}
\alias{Calc_EZ}
\title{#'@title Default parameters of ebpm
#'@export
ebpm_control_default = function(){
  list(pi0 = 'estimate',
       g_init = NULL,
       fix_g = FALSE,
       control =  NULL)
}}
\usage{
Calc_EZ(X, K, EZ, ql_hat, qf_hat)
}
\description{
#'@title Default parameters of smash gen
#'@param filter.number,family wavelet basis, see wavethresh pakcage for more details.
#'@export
splitting_control_default = function(){
  list(Eb_init = NULL,
       sigma2_init = NULL,
       est_sigma2 = TRUE,
       maxiter = 100,
       tol=1e-5,
       filter.number = 1,
       family = 'DaubExPhase',
       verbose=FALSE,
       printevery = 10,
       ebnm_params=list(mode=0),
       optim_method='L-BFGS-B')
}
}
