% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_MF_splitting.R
\name{calc_split_PMF_obj_flashier}
\alias{calc_split_PMF_obj_flashier}
\title{calc splitting PMF objective function.}
\usage{
calc_split_PMF_obj_flashier(
  Y,
  S,
  sigma2,
  M,
  V,
  fit_flash,
  KL_LF,
  const,
  var_type
)
}
\description{
calc splitting PMF objective function.
}
