% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_MF_splitting_low_memory.R
\name{calc_split_PMF_obj_flashier_low_memory}
\alias{calc_split_PMF_obj_flashier_low_memory}
\title{calc splitting PMF objective function, low memory.}
\usage{
calc_split_PMF_obj_flashier_low_memory(
  Y,
  S,
  sigma2,
  M,
  V,
  fit_flash,
  KL_LF,
  const,
  var_type
)
}
\description{
calc splitting PMF objective function, low memory.
}
