% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_log_controls.R
\name{check_flash_signs}
\alias{check_flash_signs}
\title{Check loadings and factors signs, to make sure they match ebnm.fn}
\usage{
check_flash_signs(flash_control)
}
\description{
Check loadings and factors signs, to make sure they match ebnm.fn
}
