% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_mix.R
\name{cluster.mix}
\alias{cluster.mix}
\title{Xing's smoothed topic model}
\usage{
cluster.mix(y, smooth = TRUE, pi0 = NULL, phi0 = NULL, K, tol, maxit,
  nugget = FALSE, nug_control = list())
}
\description{
Xing's smoothed topic model
}
