% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebbp_beta_mixture.R
\name{ebbp_beta_mixture}
\alias{ebbp_beta_mixture}
\title{Empirical Bayes for estimating Binomial probability}
\usage{
ebbp_beta_mixture(x, nx, shape = NULL, point_mass = TRUE,
  nullweight = 10, weight = rep(1, length(x)), g_init = NULL,
  fix_g = FALSE, control = NULL)
}
\arguments{
\item{x:}{a vector of the number of successes}

\item{nx:}{a vector of the number of trials.}

\item{shape:}{parameter of mixture Beta component, Beta(shape,shape).}

\item{point_mass:}{whether include delta(1/2) in the mixture component.}

\item{nullweight:}{extra amount of weight on the first mixing proportion.}

\item{weight:}{a vector of weights}

\item{control:}{controls of mixSQP}
}
\value{
prior, posterior, likelihood
}
\description{
Empirical Bayes for estimating Binomial probability
}
