% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_smooth.R
\name{ebnm_dwt}
\alias{ebnm_dwt}
\title{Empirical Bayes wavelet smoothing DWT wrapper for flashier}
\usage{
ebnm_dwt(x, s, g_init, fix_g, filter.number = 1, family = "DaubExPhase")
}
\arguments{
\item{x}{data}

\item{s}{known standard error}

\item{g_init}{a list of priors for each scale}

\item{fix_g}{whether fix prior}
}
\value{
a list of
 \item{posterior:}{posterior mean and 2nd moment}
 \item{fitted_g:}{estimated prior}
 \item{log_likelihood:}{log likelihood}
}
\description{
Smooth homogeneous Gaussian data.
}
