% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_smooth.R
\name{ebnm_dwt_haar}
\alias{ebnm_dwt_haar}
\title{Empirical Bayes wavelet smoothing haar wavelet for flashier}
\usage{
ebnm_dwt_haar(x, s, g_init = NULL, fix_g = FALSE, output)
}
\description{
Empirical Bayes wavelet smoothing haar wavelet for flashier
}
