% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_smooth.R
\name{ebnm_dwt_symlet}
\alias{ebnm_dwt_symlet}
\title{Empirical Bayes wavelet smoothing symlet wavelet for flashier}
\usage{
ebnm_dwt_symlet(x, s, g_init = NULL, fix_g = FALSE, output)
}
\description{
Empirical Bayes wavelet smoothing symlet wavelet for flashier
}
