% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_smooth.R
\name{ebnm_ndwt}
\alias{ebnm_ndwt}
\title{This function is an EBNM function that fits ndwt method}
\usage{
ebnm_ndwt(x, s, g_init, fix_g, output)
}
\description{
This function is an EBNM function that fits ndwt method
}
