% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm.R
\name{ebpm_control_default}
\alias{ebpm_control_default}
\title{Default parameters of ebpm}
\usage{
ebpm_control_default()
}
\arguments{
\item{shape:}{shape parameters in to `ebpm` function, default to be 1}

\item{point_mass:}{whether to use point mass estiamte.}

\item{nullweight:}{weight on the first mixing component}

\item{scale:}{scale parameters to ebpm}
}
\description{
Default parameters of ebpm
}
