% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpm_exp_mixture.R
\name{ebpm_exp_mixture}
\alias{ebpm_exp_mixture}
\title{Empirical Bayes Poisson Mean}
\usage{
ebpm_exp_mixture(x, s = 1, scale = c(1e-10, 1e-05, 0.01, 0.1, 1, 10,
  100, 1e+05), point_mass = F, nullweight = 1000, weight = rep(1,
  length(x)), g_init = NULL, fix_g = FALSE, m = 2, control = NULL,
  low = NULL, d = NULL, shape = 1)
}
\description{
Modified from Zihao Wang's ebpm package
}
