% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf.R
\name{ebpmf}
\alias{ebpmf}
\title{Empirical Bayes Poisson Matrix Factorization}
\usage{
ebpmf(X, K, pm_func = ebpm::ebpm_point_gamma, init = list(qg = NULL,
  init_method = "scd", init_iter = 20), pm_control = NULL,
  fix_g = list(l = FALSE, f = FALSE), maxiter = 100, tol = 1e-08,
  verbose = FALSE)
}
\arguments{
\item{X}{count matrix (dim(X) = c(n, p)).}

\item{pm_func}{function for solving the \code{ebpm} subproblem; can be
\code{ebpm_point_gamma, ebpm_two_gamma, ebpm_exponential_mixture, ebpm_gamma_mixture_single_scale}}

\item{init}{list(qg, init_method, init_iter)}

\item{pm_control}{control parameters for pm_func function}

\item{fix_g}{list(l, f) where l, f are either TRUE or FALSE}

\item{maxiter}{maximum number of iterations}

\item{tol}{stopping tolerance for ELBO}

\item{verbose}{print progress if set TRUE}

\item{k}{number of topics}
}
\value{
A list containing elements:
    \describe{
      \item{\code{ql}}{approximate posterior for l}
      \item{\code{gl}}{fitted g for l}
      \item{\code{kl_l}}{kl divergence between q and g for l}
      \item{\code{qf}}{approximate posterior for f}
      \item{\code{gf}}{fitted g for f}
      \item{\code{kl_f}}{kl divergence between q and g for f}
     }
}
\description{
Empirical Bayes Poisson Matrix Factorization
}
\examples{
To add
}
