% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_bg.R
\name{ebpmf_bg}
\alias{ebpmf_bg}
\title{Empirical Bayes Poisson Matrix Factorization (Background Model)}
\usage{
ebpmf_bg(X, K, pm_func = list(f = ebpm::ebpm_gamma_mixture, l = mle_pm),
  init = list(qg = NULL, init_method = "scd", init_iter = 20),
  pm_control = NULL, fix_g = list(l = FALSE, f = FALSE),
  maxiter = 100, tol = 1e-08, verbose = FALSE)
}
\arguments{
\item{X}{count matrix (dim(X) = c(n, p)).}

\item{pm_func}{function for solving the \code{ebpm} subproblem; can be
\code{ebpm_point_gamma, ebpm_two_gamma, ebpm_exponential_mixture, ebpm_gamma_mixture_single_scale}}

\item{init}{list(qg, init_method, init_iter)}

\item{pm_control}{control parameters for pm_func function}

\item{fix_g}{list(l, f) where l, f are either TRUE or FALSE}

\item{maxiter}{maximum number of iterations}

\item{tol}{stopping tolerance for ELBO}

\item{verbose}{print progress if set TRUE}

\item{k}{number of topics}
}
\value{
A list containing elements:
    \describe{
      \item{\code{l0}}{sample-wise mean}
      \item{\code{f0}}{feature-wise mean}
      \item{\code{qg}}{approximate posterior for l}
      \item{\code{ELBO}}{ELBO objective for this VEB algorithm}
     }
}
\description{
Empirical Bayes Poisson Matrix Factorization (Background Model)
}
\examples{
To add
}
