% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_identity.R
\name{ebpmf_identity}
\alias{ebpmf_identity}
\title{Smoothed Poisson Topic Model}
\usage{
ebpmf_identity(
  X,
  K,
  init = "fasttopics",
  maxiter = 100,
  maxiter_init = 30,
  tol = 0.001,
  ebpm.fn = c(ebpm::ebpm_point_gamma, smashrgen::pois_smooth_split),
  fix_F = FALSE,
  smooth_l = FALSE,
  smooth_f = TRUE,
  smooth_control = list(),
  printevery = 10,
  verbose = TRUE,
  convergence_criteria = "mKLabs"
)
}
\arguments{
\item{X}{count matrix}

\item{K}{number of factors/ranks}

\item{init}{initialization methods, 'lee','scd' from package NNLM, or 'uniform' randomly initialize; or provide init as a list with L_init and F_init.}

\item{maxiter}{maximum iterations}

\item{tol}{stop criteria}

\item{ebpm.fn}{specify functions to use for solving the poisson subproblems}

\item{fix_F}{if TRUE, F will not be updated.}

\item{smooth_l, smooth_f}{whether smooth l or f, must match the functions in ebpm.fn}

\item{init_loss}{loss function of the initialization method, either mkl or mse.}

\item{warm_start}{whether warm starts the ebpm. Useful for when smooth_l or f is true.}

\item{return_all}{whether return all outputs or simplified ones}
}
\value{
EL,EF: posterior of loadings and factors
}
\description{
This function fits Poisson Topic Model with smooth Loading or Factors
}
\examples{
set.seed(123)
n = 120
p = 256
K= 3
L = matrix(0, nrow=n, ncol=K)
FF = matrix(0, nrow=K, ncol=p)
L[1:(n/3),1] = 1
L[((n/3)+1):(2*n/3),2] = 1
L[((2*n/3)+1):n,3] = 1
L = L + matrix(runif(n*K,0,0.5),nrow=n)
FF[1,1:(p/3)] = 1+10
FF[2,((p/3)+1):(2*p/3)] = 1+10
FF[3,((2*p/3)+1):p] = 1+10
lambda = L \%*\% FF
X = matrix(rpois(n=length(lambda),lambda),nrow=n)
image(X)
}
