% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_identity_init.R
\name{ebpmf_identity_init}
\alias{ebpmf_identity_init}
\title{initialize the ebpmf identity model}
\usage{
ebpmf_identity_init(X, K, init, maxiter_init = 50, lib_size)
}
\arguments{
\item{X}{input data matrix}

\item{K}{number of topics}

\item{init}{init methods, or a list of init L and F}
}
\description{
initialize the ebpmf identity model
}
