% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_log_flash_init.R
\name{ebpmf_log_flash_init}
\alias{ebpmf_log_flash_init}
\title{the init fit of flash in ebpmf log function}
\usage{
ebpmf_log_flash_init(
  M,
  sigma2,
  l0,
  f0,
  ones_n,
  ones_p,
  loadings_sign,
  factors_sign,
  ebnm_fn,
  ebnm_fn.offset,
  S_dim,
  verbose_flash,
  fix_l0,
  fix_f0,
  Kmax,
  add_greedy_extrapolate,
  maxiter_backfitting,
  backfit_extrapolate,
  backfit_warmstart,
  init_fn.flash,
  no_backfit_kset,
  n_refit_max,
  deal_with_no_init_factor,
  var_type,
  flash_est_sigma2
)
}
\description{
the init fit of flash in ebpmf log function
}
