% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_log_flash_update.R
\name{ebpmf_log_flash_update}
\alias{ebpmf_log_flash_update}
\title{this function fits flash during ebpmf log iterations}
\usage{
ebpmf_log_flash_update(
  fit_flash,
  sigma2,
  ones_n,
  ones_p,
  iter,
  loadings_sign,
  factors_sign,
  ebnm.fn,
  ebnm.fn.offset,
  S.dim,
  verbose_flash,
  fix_l0,
  fix_f0,
  Kmax,
  add_greedy_extrapolate,
  maxiter_backfitting,
  add_greedy_every,
  add_greedy_Kmax,
  add_greedy_warmstart,
  init.fn.flash,
  no_backfit_kset
)
}
\arguments{
\item{fit_flash}{An object from running flash()}
}
\description{
this function fits flash during ebpmf log iterations
}
