% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_log_init.R
\name{ebpmf_log_init}
\alias{ebpmf_log_init}
\title{this function initializes the ebpmf with log link}
\usage{
ebpmf_log_init(
  Y,
  l0,
  f0,
  sigma2,
  var_type,
  M_init,
  verbose,
  n_cores,
  init_tol,
  printevery
)
}
\description{
sigma2, M needs initialization.
}
