% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_log.R
\name{ebpmf_log_update_sigma2}
\alias{ebpmf_log_update_sigma2}
\title{update sigma2}
\usage{
ebpmf_log_update_sigma2(
  fit_flash,
  sigma2,
  v_sum,
  var_type,
  cap_var_mean_ratio,
  a0,
  b0,
  n,
  p
)
}
\description{
update sigma2
}
