% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_point_gamma.R
\name{ebpmf_point_gamma}
\alias{ebpmf_point_gamma}
\title{Empirical Bayes Poisson Matrix Factorization with Prior Family Mixture of Exponential}
\usage{
ebpmf_point_gamma(X, K, maxiter.out = 10, maxiter.int = 1,
  verbose = F, seed = 123)
}
\arguments{
\item{X}{count matrix (dim(X) = c(n, p)).}

\item{K}{number of topics}

\item{maxiter.out}{maximum iterations in the outer  loop}

\item{maxiter.int}{maximum iterations in the inner loop}

\item{verbose}{T if print ELBO}

\item{seed}{random seed}
}
\value{
A list containing elements (this is freeuqntly updated):
    \describe{
      \item{\code{qls_mean}}{A n by k matrix: Approximate posterior mean for L}
      \item{\code{qls_mean_log}}{A n by k matrix: Approximate posterior log mean for L}
      \item{\code{gls}}{A list of K elements, each element is the estimated prior for the kth column of L}
      \item{\code{qfs_mean}}{A p by k matrix: Approximate posterior mean for F}
      \item{\code{qfs_mean_log}}{A p by k matrix: Approximate posterior log mean for F}
      \item{\code{gfs}}{A list of K elements, each element is the estimated prior for the kth column of F}
     }
}
\description{
Uses Empirical Bayes to fit the model \deqn{X_{ij}  ~ Poi(\sum_k L_{ik} F_{jk})} with \deqn{L_{.k} ~ g_k()}, where dim(X) = c(n,p); dim(L) = c(n,K); dim(F) = c(p,K)
with g_k being either Mixture of Exponential, or Point Gamma
}
\details{
The model is fit in 2 stages: i) estimate \eqn{g} by maximum likelihood (over pi_k)
ii) Compute posterior distributions for \eqn{\lambda_j} given \eqn{x_j,\hat{g}}.
}
\examples{
To add

}
