% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_log_controls.R
\name{flash_extra_control}
\alias{flash_extra_control}
\title{This function returns init.fn.flash and no_backfit_kset}
\usage{
flash_extra_control(loadings_sign, factors_sign, fix_l0, fix_f0)
}
\description{
This function returns init.fn.flash and no_backfit_kset
}
