% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm.R
\name{init_stm}
\alias{init_stm}
\title{initialize the stm model}
\usage{
init_stm(X, K, init, init_loss)
}
\arguments{
\item{X}{input data matrix}

\item{K}{number of topics}

\item{init}{init methods, or a list of init L and F}

\item{init_loss}{mkl or mse}
}
\description{
initialize the stm model
}
