% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_HALS.R
\name{NMF_HALS}
\alias{NMF_HALS}
\title{HALS algorithm for NMF}
\usage{
NMF_HALS(X, k, smooth_u = F, smooth_v = T, ord = 1,
  smooth_method = "trendfiltering", filter.number = 1,
  family = "DaubExPhase", pos_u, pos_v, maxiter = 100,
  printevery = 10, tol = 0.01)
}
\arguments{
\item{X}{N by p matrix}

\item{k}{number of topics}

\item{smooth_method}{'trendfiltering', 'wavelet' or 'runmed'}

\item{pos_u, pos_v}{position of observations}
}
\description{
HALS algorithm for NMF
}
