% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm.R
\name{nug_control_default}
\alias{nug_control_default}
\title{Default parameters of smash gen}
\usage{
nug_control_default()
}
\arguments{
\item{filter.number, family}{wavelet basis, see wavethresh pakcage for more details.}
}
\description{
Default parameters of smash gen
}
