% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf.R
\name{pmf}
\alias{pmf}
\title{Poisson Matrix Factorization}
\usage{
pmf(X, K, init, maxiter = 100, verbose = FALSE, seed = 123)
}
\arguments{
\item{X}{count matrix (dim(X) = c(n, p)).}

\item{K}{number of topics}

\item{init}{Either NULL or \code{list(L, F)}}

\item{seed}{only used when init is NULL}
}
\value{
\code{list(L, F, log_liks)}
}
\description{
Poisson Matrix Factorization
}
