% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_bg.R
\name{pmf_bg}
\alias{pmf_bg}
\title{Poisson Matrix Factorization with background}
\usage{
pmf_bg(X, K, init, fix_option, maxiter = 100, verbose = FALSE, seed = 123)
}
\arguments{
\item{X}{count matrix (dim(X) = c(n, p)).}

\item{K}{number of topics}

\item{init}{Either NULL or \code{list(l0, f0, L, F)}}

\item{seed}{only used when init is NULL}
}
\value{
\code{list(l0, f0, L, F, log_liks)}
}
\description{
Poisson Matrix Factorization with background
}
