% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_greedy.R
\name{pmf_greedy}
\alias{pmf_greedy}
\title{Poisson Matrix Factorization (greedy)}
\usage{
pmf_greedy(X, K, init, maxiter = 100, rate = 0.9, verbose = FALSE, seed = 123)
}
\arguments{
\item{X}{count matrix (dim(X) = c(n, p)).}

\item{K}{number of topics}

\item{init}{Either NULL or \code{list(L, F)}}

\item{maxiter}{for each k}

\item{seed}{only used when init is NULL}
}
\value{
\code{list(L, F, log_liks)}
}
\description{
Poisson Matrix Factorization (greedy)
}
