% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_subject.R
\name{pmf_subject}
\alias{pmf_subject}
\title{Poisson Matrix Factorization (using subject information)}
\usage{
pmf_subject(X, u, K, init = NULL, maxiter = 100, verbose = FALSE, seed = 123)
}
\arguments{
\item{X}{count matrix (dim(X) = c(n, p)).}

\item{u}{vector encoding subject (length(u) = n; elements need to be from {1,..., length(unique(u))}).}

\item{K}{number of topics}

\item{init}{Either NULL or \code{list(L, F)}}

\item{seed}{only used when init is NULL}
}
\value{
\code{list(L, F, log_liks)}
}
\description{
Poisson Matrix Factorization (using subject information)
}
