% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson2multinom.R
\name{poisson_to_multinom}
\alias{poisson_to_multinom}
\title{standard loadings and factors from Poisson matrix Factorization}
\usage{
poisson_to_multinom(FF, L)
}
\arguments{
\item{L:}{n by k matrix}

\item{FF:}{k by p matrix}
}
\description{
standard loadings and factors from Poisson matrix Factorization
}
