% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_bg_util.R
\name{rank1_bg}
\alias{rank1_bg}
\title{Empirical Bayes Poisson Matrix Factorization, Background Model (rank 1)}
\usage{
rank1_bg(
  d,
  X_rs,
  X_cs,
  l0,
  f0,
  pm_func,
  pm_control,
  ql,
  gl,
  kl_l,
  qf,
  gf,
  kl_f,
  fix_option
)
}
\description{
Empirical Bayes Poisson Matrix Factorization, Background Model (rank 1)
}
