% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_wbg2_subject_util.R
\name{rank1_wbg2_subject}
\alias{rank1_wbg2_subject}
\title{Empirical Bayes Poisson Matrix Factorization, Background Model (rank 1)}
\usage{
rank1_wbg2_subject(
  d,
  X_rs,
  X_cs,
  S,
  u,
  m,
  w_log_k,
  pm_func,
  pm_control,
  ql,
  gl,
  kl_l,
  qf,
  gf,
  kl_f,
  fix_option
)
}
\description{
Empirical Bayes Poisson Matrix Factorization, Background Model (rank 1)
}
