% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_poisson_matrix.R
\name{sim_data_log}
\alias{sim_data_log}
\title{simulate Poisson matrix}
\usage{
sim_data_log(
  n,
  p,
  K = 3,
  d = NULL,
  add_background = TRUE,
  background_unif_range = c(1, 2),
  pi0 = 0.8,
  var1 = 1,
  var_e = 0,
  n_simu = 10,
  seed = 12345
)
}
\description{
simulate Poisson matrix
}
\details{
The DGP is
\deqn{Y\sim Poisson(\exp(l_0f_0'+LF'+E)),}
\deqn{l_{0i}\sim Uniform(),f_{0j}\sim Uniform(),}
\deqn{l_{ki}\sim N(0,d_k),f_{kj}\sim \pi_0\delta_0 + (1-\pi_0)N(0,\sigma^2_1),}
\deqn{e_{ij}\sim N(0,\sigma^2).}
}
