% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_poisson_matrix.R
\name{sim_data_log_simple}
\alias{sim_data_log_simple}
\title{simulate simple Poisson matrix}
\usage{
sim_data_log_simple(N, p, K = 2, l = 20, d = NULL, var_e = 0, phi = 1.3)
}
\description{
simulate simple Poisson matrix
}
