% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_poisson_matrix.R
\name{sim_data_real}
\alias{sim_data_real}
\title{simulate Poisson matrix based on real data output}
\usage{
sim_data_real(S, L, FF, Sigma2, n_simu = 10, seed = 12345)
}
\description{
simulate Poisson matrix based on real data output
}
\details{
The DGP is
\deqn{Y\sim Poisson(S*\exp(LF'+E))}
}
