% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMF_simu_study.R
\name{simu_study_PMF}
\alias{simu_study_PMF}
\title{a simulation study function}
\usage{
simu_study_PMF(
  simdata,
  n_cores = 1,
  method_list = c("flash", "splitting"),
  Kmax = 10,
  var_type = "by_row",
  maxiter = 1000,
  tol = 0.1
)
}
\description{
a simulation study function
}
