% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash_gen_binom.R
\name{smash.gen.binom}
\alias{smash.gen.binom}
\title{Smooth Binomial data, accounting for nugget effect}
\usage{
smash.gen.binom(x, size, sigma = NULL, filter.number = 1,
  family = "DaubExPhase", approx_p = FALSE)
}
\arguments{
\item{filter.number, family:}{wavelet basis, see wavethresh pakcage for more details.}

\item{x:}{a vector of observed number of successes}

\item{size:}{a vector of number of trials}

\item{sigma:}{nugget effect}

\item{maxiter:}{maximum iterations, in general, 1 iteration is enough.}

\item{tol:}{tolerance to stop iterations.}

\item{approx_p:}{whether approxmate p by logistic(mu) or using numerical integration to calulcate mean of logit normal distribution.}
}
\description{
Smooth Binomial data, accounting for nugget effect
}
