% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash_gen_poiss.R
\name{smash.gen.poiss}
\alias{smash.gen.poiss}
\title{Smooth Poisson data, accounting for nugget effect}
\usage{
smash.gen.poiss(x, sigma = NULL, s = 1, method = "ti.thresh",
  ash.pm = FALSE, eps = 0.01, filter.number = 1,
  family = "DaubExPhase")
}
\arguments{
\item{filter.number, family:}{wavelet basis, see wavethresh pakcage for more details.}

\item{x:}{observed Poisson sequence}

\item{sigma:}{nugget effect}

\item{s:}{Scale factor for Poisson observations: y~Pois(scale*lambda).}

\item{method:}{smoothiing method, smash or ti.thresh. When n is large, ti.thresh is much faster.}

\item{ash.pm:}{whehter use ash posterior mean approxiamtion if x=0. If not x = x+eps}

\item{eps:}{if x=0, x = x + eps}

\item{maxiter:}{maximum iterations, in general, 1 iteration is enough.}

\item{tol:}{tolerance to stop iterations.}
}
\description{
Smooth Poisson data, accounting for nugget effect
}
