% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_PMF_flashr.R
\name{splitting_PMF_flashr}
\alias{splitting_PMF_flashr}
\title{Fit Sparse Poisson matrix factorization}
\usage{
splitting_PMF_flashr(
  Y,
  S,
  sigma2 = NULL,
  est_sigma2 = TRUE,
  Kmax = 10,
  var_type = "by_row",
  M_init = NULL,
  maxiter = 100,
  tol = 0.01,
  verbose_flash = FALSE,
  printevery = 10
)
}
\arguments{
\item{Y}{count data matrix}

\item{S}{The known scaling factor matrix, background frequency.}

\item{sigma2}{the variance term}

\item{est_sigma2}{whether estimate the variance term or fix it}
}
\value{
fitted object
}
\description{
Fit Sparse Poisson matrix factorization
}
