% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_MF_splitting_init.R
\name{splitting_PMF_init}
\alias{splitting_PMF_init}
\title{this function initializes the splitting PMF}
\usage{
splitting_PMF_init(
  Y,
  S,
  sigma2,
  var_type,
  M_init,
  verbose,
  n_cores,
  init_tol,
  printevery
)
}
\description{
sigma2, M needs initialization.
}
