% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmf.R
\name{spmf}
\alias{spmf}
\title{Poisson MF but with smoothed factors using smash.pois}
\usage{
spmf(
  X,
  K,
  L_init = NULL,
  F_init = NULL,
  smooth_F = TRUE,
  tol = 0.001,
  maxiter = 50,
  verbose = TRUE,
  printevery = 10
)
}
\description{
This is similar to Tom's sgom, but based on Poisson likelihood.
}
