% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm.R
\name{stm}
\alias{stm}
\title{Smoothed Poisson Matrix Factorization}
\usage{
stm(
  X,
  K,
  init = "scd",
  init_loss = "mkl",
  maxiter = 100,
  tol = 0.001,
  fix_F = FALSE,
  bmsm_control_l = list(),
  bmsm_control_f = list(),
  nug_control_l = list(),
  nug_control_f = list(),
  ebpm_method = "point_gamma",
  ebpm_control_l = list(),
  ebpm_control_f = list(),
  smooth_f = TRUE,
  smooth_l = FALSE,
  nugget = FALSE,
  printevery = 10
)
}
\arguments{
\item{X}{count matrix}

\item{K}{number of factors/ranks}

\item{init}{initialization methods, 'lee','scd' from package NNLM, or 'uniform' randomly initialize; or provide init as a list with L_init and F_init.}

\item{init_loss}{loss function of the initialization method, either mkl or mse.}

\item{maxiter}{maximum iterations}

\item{tol}{stop criteria}

\item{fix_F}{if TRUE, F will no be updated.}

\item{ebpm_method}{point_gamma or two_gamma}

\item{smooth_f, smooth_l}{whether to get smooth estimate of loadings or factors.}

\item{nugget}{whether to assume nugget effects}

\item{bmsm_control}{control parameters of BMSM, see bmsm_control_default()}

\item{ebpm_control}{control parameters of ebpm, see ebpm_control_default()}

\item{nug_control}{control parameters of smashgen, see nug_control_default()}

\item{return_all}{whether return all outputs or simplified ones}
}
\value{
EL,EF: posterior of loadings and factors
}
\description{
Smoothed Poisson Matrix Factorization/ smoothed topic model
}
