% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_identity.R
\name{stm_update_rank1}
\alias{stm_update_rank1}
\title{rank 1 update of the model}
\usage{
stm_update_rank1(
  l_seq,
  f_seq,
  k,
  ebpm.fn.l,
  ebpm.fn.f,
  res,
  fix_F,
  smooth_F,
  ebps_control
)
}
\description{
rank 1 update of the model
}
