% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_MF_splitting_low_memory.R
\name{sum_lfactorial_sparseMat}
\alias{sum_lfactorial_sparseMat}
\title{calculate sum of log factorial of all elements of a sparse Matrix}
\usage{
sum_lfactorial_sparseMat(Y)
}
\description{
calculate sum of log factorial of all elements of a sparse Matrix
}
