% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_MF_splitting.R
\name{vga_pois_solver_mat}
\alias{vga_pois_solver_mat}
\title{a matrix version of the solver}
\usage{
vga_pois_solver_mat(init_Val, X, S, Beta, Sigma2, maxiter = 1000, tol = 1e-08)
}
\description{
a matrix version of the solver
}
