% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vga_solver_mat.R
\name{vga_pois_solver_mat_newton_fixed_iter}
\alias{vga_pois_solver_mat_newton_fixed_iter}
\title{vga solver for fixed iteration, not necessary to convergence}
\usage{
vga_pois_solver_mat_newton_fixed_iter(M, V, Y, S, Beta, Sigma2, maxiter = 1000)
}
\description{
vga solver for fixed iteration, not necessary to convergence
}
