% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_MF_splitting.R
\name{vga_pois_solver_mat_newton_iter}
\alias{vga_pois_solver_mat_newton_iter}
\title{a matrix version of the vga solver using Newton's method, for only one/2/3/... iteration}
\usage{
vga_pois_solver_mat_newton_iter(
  M,
  X,
  S,
  Beta,
  Sigma2,
  maxiter = 1,
  tol = 1e-08
)
}
\description{
a matrix version of the vga solver using Newton's method, for only one/2/3/... iteration
}
