% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vga_solver_mat.R
\name{vga_pois_solver_mat_newton_low_memory}
\alias{vga_pois_solver_mat_newton_low_memory}
\title{a matrix version of the vga solver using Newton's method, low memory mode by partitioning the rows}
\usage{
vga_pois_solver_mat_newton_low_memory(
  M,
  Y,
  l0,
  f0,
  EL,
  EF,
  sigma2,
  var_type,
  maxiter = 1000,
  tol = 1e-08
)
}
\description{
a matrix version of the vga solver using Newton's method, low memory mode by partitioning the rows
}
