% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_splitting.R
\name{ebpmf}
\alias{ebpmf}
\title{Fit Sparse Poisson matrix factorization via variational empirical Bayes}
\usage{
ebpmf(
  Y,
  S = NULL,
  sigma2 = NULL,
  est_sigma2 = TRUE,
  ebnm.fn = ebnm::ebnm_point_normal,
  loadings_sign = 0,
  factors_sign = 0,
  Kmax = 50,
  var_type = "by_col",
  M_init = NULL,
  maxiter = 100,
  tol = 1e-05,
  maxiter_backfitting = 1,
  verbose_flash = 0,
  printevery = 10,
  verbose = FALSE,
  n_cores = 1,
  save_fit_every = Inf,
  save_fit_path = NULL,
  save_fit_name = NULL
)
}
\arguments{
\item{Y}{count data matrix}

\item{S}{The known scaling factor matrix, background frequency, default = 11'.}

\item{sigma2}{the variance term}

\item{est_sigma2}{whether estimate the variance term or fix it}

\item{ebnm.fn}{see `?flash`.}
}
\value{
fitted object
}
\description{
Fit Sparse Poisson matrix factorization via variational empirical Bayes
}
