% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_bg.R
\name{ebpmf_bg}
\alias{ebpmf_bg}
\title{Empirical Bayes Poisson Matrix Factorization (Background Model)}
\usage{
ebpmf_bg(X, K, pm_func = list(f = ebpm::ebpm_gamma_mixture, l =
  ebpm::ebpm_gamma_mixture), init = NULL, pm_control = NULL,
  fix_option = list(l0 = FALSE, f0 = FALSE, gl = FALSE, ql = FALSE, gf =
  FALSE, qf = FALSE), maxiter = 100, tol = 1e-08, verbose = FALSE,
  seed = 123)
}
\arguments{
\item{X}{count matrix (dim(X) = c(n, p)).}

\item{pm_func}{functions for solving the \code{ebpm} subproblem for \code{L} and \code{F}; 
It is a list \code{list(l, f)};
For our purpose we use `mle_pm` or `ebpm_point_gammma`for \code{L}, and \code{ebpm_gamma_mixture} for \code{F}}

\item{init}{Either \code{NULL} or \code{list(qg, l0, f0)}}

\item{pm_control}{control parameters for pm_func function}

\item{maxiter}{maximum number of iterations}

\item{tol}{stopping tolerance for ELBO}

\item{seed}{used when init is NULL}

\item{k}{number of topics}

\item{fix_g}{list(l, f) where l, f are either TRUE or FALSE}
}
\value{
A list containing elements:
    \describe{
      \item{\code{l0}}{sample-wise mean}
      \item{\code{f0}}{feature-wise mean}
      \item{\code{qg}}{list(ql, gl,qf, gf)}
      \item{\code{ELBO}}{ELBO objective for this VEB algorithm}
     }
}
\description{
Empirical Bayes Poisson Matrix Factorization (Background Model)
}
\examples{
To add
}
