% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_identity.R
\name{ebpmf_identity}
\alias{ebpmf_identity}
\title{Smoothed Poisson Topic Model}
\usage{
ebpmf_identity(
  X,
  K,
  S = NULL,
  init = "scd",
  init_loss = "mkl",
  maxiter = 100,
  tol = 1e-08,
  ebpm.fn = ebpm::ebpm_point_gamma,
  fix_F = FALSE,
  printevery = 10,
  verbose = TRUE,
  convergence_criteria = "Labs"
)
}
\arguments{
\item{X}{count matrix}

\item{K}{number of factors/ranks}

\item{S}{background}

\item{init}{initialization methods, 'lee','scd' from package NNLM, or 'uniform' randomly initialize; or provide init as a list with L_init and F_init.}

\item{init_loss}{loss function of the initialization method, either mkl or mse.}

\item{maxiter}{maximum iterations}

\item{tol}{stop criteria}

\item{fix_F}{if TRUE, F will not be updated.}

\item{smooth_method}{splitting or bmsm}

\item{bmsm_control}{control parameters of BMSM, see bmsm_control_default()}

\item{ebpm_method}{point_gamma or two_gamma}

\item{ebpm_control}{control parameters of ebpm, see ebpm_control_default()}

\item{splitting_control}{control parameters of smashgen, see splitting_control_default()}

\item{smooth_f, smooth_l}{whether to get smooth estimate of loadings or factors.}

\item{nugget}{whether to assume nugget effects}

\item{return_all}{whether return all outputs or simplified ones}
}
\value{
EL,EF: posterior of loadings and factors
}
\description{
This function fits Poisson Topic Model with smooth Loading or Factors
}
\examples{
set.seed(123)
n = 120
p = 256
K= 3
L = matrix(0, nrow=n, ncol=K)
FF = matrix(0, nrow=K, ncol=p)
L[1:(n/3),1] = 1
L[((n/3)+1):(2*n/3),2] = 1
L[((2*n/3)+1):n,3] = 1
L = L + matrix(runif(n*K,0,0.5),nrow=n)
FF[1,1:(p/3)] = 1+10
FF[2,((p/3)+1):(2*p/3)] = 1+10
FF[3,((2*p/3)+1):p] = 1+10
lambda = L \%*\% FF
X = matrix(rpois(n=length(lambda),lambda),nrow=n)
image(X)
}
