% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_log.R
\name{ebpmf_log}
\alias{ebpmf_log}
\title{Fit empirical Bayes matrix factorization}
\usage{
ebpmf_log(
  Y,
  l0 = NULL,
  f0 = NULL,
  var_type = "by_col",
  sigma2 = NULL,
  est_sigma2 = TRUE,
  M_init = NULL,
  ebnm.fn = ebnm::ebnm_point_normal,
  loadings_sign = 0,
  factors_sign = 0,
  Kmax_init = 30,
  add_greedy_Kmax = 1,
  add_greedy_warmstart = TRUE,
  add_greedy_extrapolate = FALSE,
  add_greedy_init = "new_init",
  add_greedy_every = 1,
  batch_size = 1000,
  maxiter = 100,
  maxiter_backfitting = 1,
  maxiter_vga = 3,
  conv_tol = 1e-05,
  init_tol = 1e-05,
  vga_tol = 0.001,
  verbose_flash = 0,
  printevery = 10,
  verbose = FALSE,
  n_cores = 1,
  a0 = 1,
  b0 = 1,
  cap_var_mean_ratio = 0.1,
  save_init_val = FALSE,
  return_sigma2_trace = FALSE,
  garbage_collection_every = 10
)
}
\arguments{
\item{Y}{count data matrix, in sparse format}

\item{l0, f0}{The known scaling loadings and factors.}

\item{sigma2}{the variance term}

\item{est_sigma2}{whether estimate the variance term or fix it at sigma2}

\item{ebnm.fn}{see `?flash`.}

\item{loadings_sign, factors_sign}{see `?init.fn.default`, must match ebnm.fn}

\item{Kmax_init}{the Kmax in the first flash fit}

\item{add_greedy_Kmax}{THe Kmax in add_greedy in iterations}

\item{add_greedy_warmstart, add_greedy_extrapolate}{see `?flash.add.greedy`}

\item{add_greedy_init}{either 'previous_init' or "new_init"}

\item{add_greedy_every}{perform flash add greedy every `add_greedy_every` iterations.}

\item{batch_size}{reduce memory usage for vga step by looping subsets of dataset.}

\item{maxiter, maxiter_backfitting, maxiter_vga}{max iterations for the splitting, backfitting, vga.}

\item{conv_tol, init_tol, vga_tol}{tolerance for convergence, initialization vga fit, and vga fit in iterations}

\item{a0, b0}{Inverse-Gamma(a0,b0) prior on sigma2 for regularization.}

\item{cap_var_mean_ratio}{only update sigma2 when if var/mean > (1+cap_var_mean_ratio). i.e. when overdispersion is low enough, stop updating sigma2 to boost convergence.}

\item{garbage_collection_every}{perform gc() to reduce memory usage.}
}
\value{
fitted object
}
\description{
Fit empirical Bayes matrix factorization
}
