% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_log_flash_init.R
\name{ebpmf_log_flash_init}
\alias{ebpmf_log_flash_init}
\title{the init fit of flash in ebpmf log function}
\usage{
ebpmf_log_flash_init(
  M,
  sigma2,
  l0,
  f0,
  ones_n,
  ones_p,
  loadings_sign,
  factors_sign,
  ebnm.fn,
  ebnm.fn.offset,
  S.dim,
  verbose_flash,
  fix_l0,
  fix_f0,
  Kmax,
  add_greedy_extrapolate,
  maxiter_backfitting,
  backfit_extrapolate,
  backfit_warmstart,
  init.fn.flash,
  no_backfit_kset
)
}
\description{
the init fit of flash in ebpmf log function
}
