% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash_gen_poiss.R
\name{smash.gen.poiss}
\alias{smash.gen.poiss}
\title{Smooth Poisson sequence, accounting for nugget effect}
\usage{
smash.gen.poiss(
  x,
  nugget = NULL,
  s = 1,
  transformation = "vst",
  method = "ti.thresh",
  robust = T,
  robust.q = 0.99,
  nug.init = NULL,
  ash.pm = FALSE,
  eps = "estimate",
  filter.number = 1,
  family = "DaubExPhase",
  maxiter = 10,
  tol = 0.01
)
}
\arguments{
\item{x}{observed Poisson sequence}

\item{nugget}{nugget effect}

\item{s}{Scale factor for Poisson observations: y~Pois(scale*lambda), can be a vector.}

\item{transformation}{transformation of Poisson data, either 'vst' or 'lik_expansion'; 'vst' for varaince stablizing transformation; 'lik_expansion' for likelihood expansion}

\item{method}{smoothing method for Gaussian sequence, either 'smash' or 'ti.thresh'. When n is large, ti.thresh is much faster}

\item{robust}{whether perform robust wavelet regression}

\item{robust.q}{quantile to determine outliers}

\item{nug.init}{init value of nugget effect, either a scalar or NULL}

\item{ash.pm}{If choose lik_expansion, whehter use ash posterior mean approxiamtion if x=0. If not x = x+eps.}

\item{eps}{If choose lik_expansion, if x=0, set x = x + eps. Either input a numerical value or 'estimate'. If estimate, eps = sum(x==1)/sum(x<=1)}

\item{filter.number, family}{wavelet basis, see wavethresh pakcage for more details}

\item{maxiter}{max iterations for estimating nugget effect}

\item{tol}{tolerance to stop iterations.}
}
\value{
estimated smoothed lambda, estimated nugget effect.
}
\description{
Smooth Poisson sequence, accounting for nugget effect
}
