% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_MF_splitting.R
\name{splitting_PMF_flashier}
\alias{splitting_PMF_flashier}
\title{Fit Sparse Poisson matrix factorization}
\usage{
splitting_PMF_flashier(
  Y,
  S = NULL,
  var_type = "by_col",
  sigma2 = NULL,
  est_sigma2 = TRUE,
  ebnm.fn = ebnm::ebnm_point_normal,
  loadings_sign = 0,
  factors_sign = 0,
  Kmax_init = 50,
  add_greedy_Kmax = 1,
  add_greedy_warmstart = TRUE,
  add_greedy_extrapolate = FALSE,
  add_greedy_init = "previous_init",
  add_greedy_every = 1,
  M_init = NULL,
  maxiter = 100,
  maxiter_backfitting = 1,
  maxiter_vga = 1,
  conv_tol = 1e-05,
  init_tol = 1e-05,
  vga_tol = 1e-05,
  verbose_flash = 0,
  printevery = 10,
  verbose = FALSE,
  n_cores = 1,
  save_fit_every = Inf,
  save_fit_path = NULL,
  save_fit_name = NULL
)
}
\arguments{
\item{Y}{count data matrix}

\item{S}{The known scaling factor matrix, background frequency.}

\item{sigma2}{the variance term}

\item{est_sigma2}{whether estimate the variance term or fix it at sigma2}

\item{ebnm.fn}{see `?flash`.}

\item{loadings_sign, factors_sign}{see `?init.fn.default`, must match ebnm.fn}

\item{Kmax_init}{the Kmax in the first flash fit}

\item{add_greedy_Kmax}{THe Kmax in add_greedy in iterations}

\item{add_greedy_warmstart, add_greedy_extrapolate}{see `?flash.add.greedy`}

\item{add_greedy_init}{either 'previous_init' or "new_init"}

\item{add_greedy_every}{perform flash add greedy every `add_greedy_every` iterations.}

\item{maxiter, maxiter_backfitting, maxiter_vga}{max iterations for the splitting, backfitting, vga.}

\item{conv_tol, init_tol, vga_tol}{tolerance for convergence, initialization vga fit, and vga fit in iterations}
}
\value{
fitted object
}
\description{
Fit Sparse Poisson matrix factorization
}
