% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_MF_splitting.R
\name{splitting_PMF_flashier}
\alias{splitting_PMF_flashier}
\title{Fit Sparse Poisson matrix factorization}
\usage{
splitting_PMF_flashier(
  Y,
  S,
  sigma2 = NULL,
  est_sigma2 = TRUE,
  Kmax = 10,
  var_type = "by_col",
  M_init = NULL,
  maxiter = 100,
  tol = 1e-05,
  maxiter_backfitting = 10,
  verbose_flash = 0,
  printevery = 10,
  verbose = FALSE,
  n_cores = 1,
  save_fit_every = Inf,
  save_fit_path = NULL,
  save_fit_name = NULL
)
}
\arguments{
\item{Y}{count data matrix}

\item{S}{The known scaling factor matrix, background frequency.}

\item{sigma2}{the variance term}

\item{est_sigma2}{whether estimate the variance term or fix it}
}
\value{
fitted object
}
\description{
Fit Sparse Poisson matrix factorization
}
